requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
openai
pydantic
python-dotenv
